Tried something new with send - not multithreaded
I moved the write() call to the BareSend.send function from the new
thread. I then had to add a timer to get sendDone to be after the return
to send.

It doesn't work very well. When I run RCTLB on two nodes I miss a lot of
packets when I'm sending 1 pkt/sec as well as receiving 1 pkt/sec. If I
stop the transmit half it runs smoothly. I'm not sure what the blocking
issue is.