Fixed the LPL setting commands.
LPL window is different than interval. The window is the period
of time that the radio listens for activity, the interval is the
period that the radio sleeps between active windows.

Although we don't implement the listening side of LPL (we have
infinite power), we do calculate how many times to resend the packet
using the parameters that the listening side will use.

Examining the cc2420 and rfxlink-based drivers it looks like 320uS
is a standard window length so I added it to RadioConfig. It can
be overridden at compile-time.