Better handling of return info from packet receive
I believe the two bytes that follow each packet from the read() call are
now parsed and saved correctly.

I added a ack struct to the metadata info because that seemed like the
cleanest way to keep track of whether an ack was received. Putting it in
the cc2520_metadata_t struct would violate the assumed conditions of the
cast in ReciveP (the ack field wouldn't point to meaningful data there).