New method for handling io with select
Each module that wants to read/write from a file (socket, etc.) will
register it with IOManagerC. When the tinyos application goes to "sleep"
it will block on the select call and wait for a file to come in.

The timeout on the select call will be used for timers. If no file comes
in then the timer will be called.

The reason for switching is it makes it easier to add udp sockets to the
mix without having a bunch of threads to worry about. It would be nice
to have BLIP register its bind() calls with linux so that packets coming
in from either side can reach the code in tinyos that is bound to a
given port.